requirements.txt
numpy==1.21.6
pandas==1.3.5
Pillow==9.3.0
pytesseract==0.3.10
tensorflow==2.11.0
ultralytics==8.0.141
flask==2.2.5
simple-salesforce==1.12.5
